usb: device_next: do not set alternate 0 for non-existent interfaces
A request to set alternate zero for a non-existent interface
would succeed because of the inappropriate order of checks.
Move the check if the desired alternate is the same as the
current one after the check if interface exists.

(cherry picked from commit 2dbccf065429836bc311a5deef1ba9782725a0f9)

Original-Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>
GitOrigin-RevId: 2dbccf065429836bc311a5deef1ba9782725a0f9
Change-Id: I1d8da4bd6cd052b6c062ce0e4f8cbee5c34fa1ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4526216
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>