drivers: flash: nrf_qspi_nor: Mark device as busy when locked
Mark the device as busy when it is locked to perform a flash operation
so that the PM subsystem does not try to suspend it. Also refuse to run
a PM action when the device is busy (in case `qspi_nor_pm_action()` is
called directly, not by the PM subsystem).

(cherry picked from commit 7b8170dd278ed40913d5df0944f34d81975ae8f0)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: 7b8170dd278ed40913d5df0944f34d81975ae8f0
Change-Id: Id9687e18f3853fa55c438bb846c807efa816cc56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4477350
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>