kernel: event modification functions return previous value
Update the return value of functions that modify the internal event
state from `void` to `uint32_t`, so that calling code can determine
whether the event was already in a given state, or if the call modified
it.

This simplifies the usage of `struct k_event` as an alternative to
`atomic_t` that users can block on.

Implements #57216

(cherry picked from commit 4ce1f03aa98c896713834793b3bfc6581d6ac450)

Original-Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>
GitOrigin-RevId: 4ce1f03aa98c896713834793b3bfc6581d6ac450
Change-Id: I98f1d24b0f60407d431eedbc4694d1539e248cde
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4672881
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>