cmake: linker: arm: put RAM sections in RAM region
Cleanup RAM layout so that the RAM_REGION is now in the RAM memory
region.  Put .bss in RAM_REGION.

This means we have a proper program header for the RAM region.

As part of this extend zephyr_linker_symbol to allow it to take
an OBJECT paramater to specify the region to associate a symbol to.

(cherry picked from commit f5eada5553467c732e7e5bf283e72b66d7be71a9)

Original-Signed-off-by: Kumar Gala <kumar.gala@intel.com>
GitOrigin-RevId: f5eada5553467c732e7e5bf283e72b66d7be71a9
Change-Id: Idee06cc66b97072f8f2561a8e64eceeef79bc0dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4503202
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>