nrf52_bsim: twister yaml: Reduce filtering
Two parts:

1)
The nrf52_bsim has had an overall driver filter since
the begining.
But nowadays it supports most of the nrf drivers, so
it can be used to run these drives' tests.
It is actually the only platform which allows running
these drivers in CI without HW (for ex. upstream).

Instead of filtering all drivers, filter only the
drivers which cannot yet be used in this board.

2)
This platform is not a good target to test the
modem drivers, as will not be able to run those until
somebody would want to include models of their respective HW,
for which there is currently no intention.
Build only tests are better suited in native_posix.

Moreover, most of those will fail to build due to their
use of either a GPIO or UART which are not yet supported
in this board.

(cherry picked from commit 75f37f196b15273c61d6a0ec40f58365573cb065)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 75f37f196b15273c61d6a0ec40f58365573cb065
Change-Id: I67870264fda63dcefcb0877e7bb1dbf081e87ad3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4509086
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>