drivers: flash: Add micron mt29fx Nand flashers drivers
The flash is controlled using a set of control singals - chip select,
address latch enable (ALE), command latch enable (CLE), and
ready/busy signal. A multiplexed 8/16-bit data/address carries address,
command and data.

The signals and data are controlled using a MEMC interface (i.e.
xmc4xxx memc or sam0 memc).

During an address/command phase, the ALE and CLE signals need to be
set, respectively (in addition to the chip select signal).
Often, the address lines of the MEMC interface are used to carry the
addition ALE and CLE signals.
For example address line correspoding to bit 16 of the 32-address bus
can carry ALE and bit 17 the CLE.