Bluetooth: host: Fix dangling reference in service changed
The bt_gatt_indicate() expects its parameters to remain valid while the
indicate procedure is active. But the `sc_range` variable was local to
the function. It is assigned to the `data` field and passed on to
bt_gatt_indicate(). The memory associated with `sc_range` goes out of
scope as soon as the function returns thereby breaking the contract of
the API. This dangling reference will lead to undefined behavior.
This is now fixed by making the `sc_range` array static and further
making it an array of arrays, as the sc_range may have different values
for each connection.

Found as violation of MISRA C:2012 and CERT DCL30-C by sonarcloud.

(cherry picked from commit 491628ae0614377ae64788ca3641c7698c63147a)

Original-Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>
GitOrigin-RevId: 491628ae0614377ae64788ca3641c7698c63147a
Change-Id: I1e4242bb1672d9ee844fce758176479394954789
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4526242
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>