Bluetooth: BAP: Broadcast source fix idle state
The idle state was never set, thus providing a poorly
implemented state machine. Updated the code so that unused
code was removed and so that the idle state is properly set.

Optimized the way that a state is set for all endpoints
for a broadcast source with a new function.

(cherry picked from commit c66f210f6b70011e3512831729b20d9ffc03cda5)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: c66f210f6b70011e3512831729b20d9ffc03cda5
Change-Id: I50be9c29ca2f1b0bc635071ac607b0e928885e4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4537491
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>