libc/common: Make sure malloc arena is PoT aligned on Risc-V
The malloc arena needs to be aligned to a suitable size on Risc-V, and the
Z_RISCV_STACK_GUARD_SIZE variable is a helpful proxy for what size that
is. However, that is not always a power of two, so round it up to make the
linker capable of performing the necessary allocation.

(cherry picked from commit 5279acba9d1b049210c7bc2b75a49b5097dca11b)

Original-Signed-off-by: Keith Packard <keithp@keithp.com>
GitOrigin-RevId: 5279acba9d1b049210c7bc2b75a49b5097dca11b
Change-Id: I4f37fc46ed702a1552fed665fbbba746f7588827
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4522944
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>