logging: Fix user space crash when runtime filtering is on
Logging module data (including filters) are not accessible by
the user space. Macro for creating logs where creating local
variable with filters before checking is we are in the user
context. It was not used in that case but creating variable
was violating access writes that resulted in failure.

Removing variable creation and using filters directly in the
if clause but after checking condition that it is not the
user context. With this approach data is accessed only in
the kernel mode.

(cherry picked from commit 9bf333aa8d29fad4e4fe6fac301ee627a5880cfb)

Original-Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>
GitOrigin-RevId: 9bf333aa8d29fad4e4fe6fac301ee627a5880cfb
Change-Id: I233177bbcac552d4b491a1f962e97c654dd72309
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4543680
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>