libc/common: Add memalign
Memalign is another name for the posix aligned_alloc function, although it
has weaker restrictions on the relationship between the alignment and size.

memalign() is used internally by the libstdc++ when built for 'newlib'
targets (which includes picolibc) instead of aligned_alloc() due to a bug
in gcc, so we need to provide an implementation of this when using that
library, even though it's not part of the Zephyr C library API.

When a fix for the libstdc++ is merged upstream and can be consider a
reasonable dependency for Zephyr, this work-around can be removed.

(cherry picked from commit 101cdcd86e462c3a13fc509f5d586c733308b019)

Original-Closes: #57899
Original-Signed-off-by: Keith Packard <keithp@keithp.com>
GitOrigin-RevId: 101cdcd86e462c3a13fc509f5d586c733308b019
Change-Id: Ib0965b6d958e15655531179239f543800c14b0d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4773031
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>