dts: bindings: interrupt-controller: GIC: Update description
When the GIC driver was originally introduced, it was only used on
Cortex-R SoCs. However, this is not the case anymore. Update the
description to reflect that this driver is not specific to Cortex-R.

(cherry picked from commit 88f4353ce133a6bc46026084b726f1efb366f763)

Original-Signed-off-by: Piotr Wojnarowski <pwojnarowski@antmicro.com>
GitOrigin-RevId: 88f4353ce133a6bc46026084b726f1efb366f763
Change-Id: I893e84dd4230e095605e3082ce5a330adfd587ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4625137
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>