drivers: gpio: pca95xx: add latch for interrupt pins
Basically, once a pin is configured for interrupt, we enable the latch
on it, in order not to miss short pulses.
The bottom half of the interrupt then reads the input port twice, to
read the latch and then read the potentially new state of the pins, to
compute the posedge, negedge, and level callbacks accordingly.

(cherry picked from commit 06e4ba88b417b666e345e6af8aecb804de002fc5)

Original-Signed-off-by: Vincent Geneves <vgeneves@kalray.eu>
GitOrigin-RevId: 06e4ba88b417b666e345e6af8aecb804de002fc5
Change-Id: Ic93b5f3b7e5fd666e3401f6821523589a4dd0665
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4571705
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>