riscv: prevent possible deadlock on SMP with FPU sharing
Let's consider CPU1 waiting on a spinlock already taken by CPU2.

It is possible for CPU2 to invoke the FPU and trigger an FPU exception
when the FPU context for CPU2 is not live on that CPU. If the FPU context
for the thread on CPU2 is still held in CPU1's FPU then an IPI is sent
to CPU1 asking to flush its FPU to memory.

But if CPU1 is spinning on a lock already taken by CPU2, it won't see
the pending IPI as IRQs are disabled. CPU2 won't get its FPU state
restored and won't complete the required work to release the lock.

Let's prevent this deadlock scenario by looking for a pending FPU IPI
from the arch_spin_relax() hook and honor it.

(cherry picked from commit 18ff7ae998acd5556864f7c0708b3528abce0be3)

Original-Signed-off-by: Nicolas Pitre <npitre@baylibre.com>
GitOrigin-RevId: 18ff7ae998acd5556864f7c0708b3528abce0be3
Change-Id: I3e1a690b3bc39e2cde35a4fcbd4362397ecf82e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4566195
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>