scripts: Twister ConfigurationError Fix
Current implementation of ConfigurationError causes
an exception to be thrown.
It is because a PosixPath is being concatenated with an str with a '+'
operator, which is not permitted.
The path in question is now cast to an str.
Additional test has been added to test_twister.py to track regressions.

(cherry picked from commit 6ce37948a05c63ab08da48c098d6dee699764059)

Original-Signed-off-by: Lukasz Mrugala <lukaszx.mrugala@intel.com>
GitOrigin-RevId: 6ce37948a05c63ab08da48c098d6dee699764059
Change-Id: I3a1629a1385d8bda58e856812ead5fa04757e87d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4570798
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>