Bluetooth: BAP: Unicast client optimize long notify reads
Optimize the read procedure when receiving long notifications
by storing the already notified data in the buffer if the
client instance is not already busy. If it is busy we have
no way to store it, and will just need to read the entire thing
later.

This also modifies when we reset the ATT buffer. Instead of reseting
when we want to use it, we reset it when we are done using it.
The reason for this is to avoid making state checks before calling
long_ase_read to determine whether it should reset the buffer
or not.

(cherry picked from commit 19c2bd28a4f7127364c2be3c0728dc978070455e)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 19c2bd28a4f7127364c2be3c0728dc978070455e
Change-Id: I1b8f15e0ac9cbfacc0bce11af10fd9e29cda4b91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4570817
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>