sensors: Create `sensor trig` subcommand
This PR relocates the sensor trigger sample application from the
`sensor_shell` sample to a subcommand in the actual sensor shell. The
subcommand has a UI for enabling a given trigger on a given sensor.
A built-in handler for the data_ready trigger is included that prints
the latest data to the log. Currently, only `SENSOR_TRIG_DATA_READY` is
supported but the groundwork is there to add others. Tested on a
`tdk_robokit1` board.

(cherry picked from commit dc9bd3afe3affc5f2d4737977ab3de627cb10af4)

Original-Signed-off-by: Tristan Honscheid <honscheid@google.com>
GitOrigin-RevId: dc9bd3afe3affc5f2d4737977ab3de627cb10af4
Change-Id: I9584a07f91a7501c7631155ea7d29410080207e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4572379
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>