drivers: sensors: npcx_tach: Clear stale data
The tachometer collects the data continuously setting a "data ready" bit
when it's ready. The availability bit has to be cleared before the
register is updated.
The driver also supports underflow detection, when the bit indicating it
is set the reading of "0" is returned.
The problem here is that there is that once the underflow bit is cleared
we might end up reading stale data.
To prevent that clear the "data ready" bit when underflow is detected

(cherry picked from commit 057888836e58b719ceaf5338b033ada497db8058)

Original-Signed-off-by: Kornel Dulęba <mindal@semihalf.com>
GitOrigin-RevId: 057888836e58b719ceaf5338b033ada497db8058
Change-Id: I487247e88c2081ea7fe68f61dc1cda7e8c0b6217
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4576988
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Kornel Dulęba <korneld@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>