Bluetooth: Host: Fix GATT Store on Write
When CCC or CF store on write is disabled and the other one is enabled,
leading to the delayed store being enabled. This cause the value of the
corresponding `n` selected option to not be stored at all.

To fix this, the checks to know if we need to store the CCC or the CF,
happening at the disconnection, are now ensuring that the CCC or the CF
are stored even if the one of them does not enable store on write.

(cherry picked from commit 7d9fe3d0b1a1313f1a7650c90b29de86d6d781eb)

Original-Signed-off-by: Théo Battrel <theo.battrel@nordicsemi.no>
GitOrigin-RevId: 7d9fe3d0b1a1313f1a7650c90b29de86d6d781eb
Change-Id: I7ac762d1fb5e2a2aa8454f54d8515cb4b8f1354f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4593589
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>