drivers: pwm: stm32: support capturing on four channels
Previously the pwm capture only support capturing on channel 1
and 2, because the slave mode controller was used, which uses
the signal TIxFP which is not available for channel 3 and 4.

This commit adds optional support for four channel capturing by
changing the method of capturing PWM signals to not use the
slave mode controller to reset the counter register. Instead the
counter is reset in the ISR. This will result in a slight loss
of accuracy but is still within an acceptable range.

(cherry picked from commit 1732651062a90486a430744d1864489413ead60f)

Original-Co-authored-by: Jeroen van Dooren <jeroen.van.dooren@nobleo.nl>
Original-Signed-off-by: Hein Wessels <heinwessels93@gmail.com>
GitOrigin-RevId: 1732651062a90486a430744d1864489413ead60f
Change-Id: Ia815b93320e0eeb4c723bd13754735cf71fdeecd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4851946
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>