qemu_cortex_m0: Fix timer pre-scaler
It's long forgotten what is the issue the comment in this FIXME is
referring to, but the clock is currently running at twice the frequency
on the qemu_cortex_m0.

This is easily provable comparing the k_msleep() delay to the wall
clock. For example:

  while (1) {
    k_msleep(1000);
    printk("hello\n");
  }

is printing "hello" at 2Hz instead of 1Hz when running with:

  qemu-system-arm -cpu cortex-m0 -machine microbit -nographic -kernel
      build/zephyr/zephyr.elf

Fix this by setting the correct pre-scaler value.

(cherry picked from commit 93fa8f84d3165a8cb82938aa6d6d0fbeef7ed8e1)

Original-Signed-off-by: Carlo Caione <ccaione@baylibre.com>
Original-Suggested-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: 93fa8f84d3165a8cb82938aa6d6d0fbeef7ed8e1
Change-Id: Iab46bab98864a4e1b0cd57bc9c4aae709f095787
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4593594
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>